Sonarqube issues

+++ anchor-image-bean/src/main/java/org/anchoranalysis/image/bean/threshold/relation/BinaryHighVoxels.java
@@ -22,7 +22,7 @@ public class BinaryHighVoxels extends RelationToThreshold {
 	
 	@Override
 	public double threshold() {
-		return BinaryValues.getDefault().getOnInt() - 1;
+		return (double) (BinaryValues.getDefault().getOnInt() - 1);
 	}
 
 	@Override
