+++ anchor-image-bean/src/main/java/org/anchoranalysis/image/bean/threshold/relation/BinaryLowVoxels.java
@@ -22,7 +22,7 @@ public class BinaryLowVoxels extends RelationToThreshold {
 	
 	@Override
 	public double threshold() {
-		return BinaryValues.getDefault().getOffInt() + 1;
+		return (double) (BinaryValues.getDefault().getOffInt() + 1);
 	}
 
 	@Override
